+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ### Updated
 
+- `io.gitlab.arturbosch.detekt:detekt-gradle-plugin -> 1.19.0`
 - `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.17`
 
 ## [0.1.0-rc.17] - 2021-11-28